2009-07-10 18:32 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * include/hbgtcore.h   * include/hbapigt.h   * source/rtl/gtapi.c   * source/rtl/hbgtcore.c   * contrib/hbct/ctwin.c   * contrib/hbct/screen1.c   * contrib/hbct/ctwfunc.c     * GetClearChar/SetClearChar changed to use USHORT for char       (was int). This resolves all remaining MSVC warnings.     * _HB_GT_BASE structure adapted to latest changes, three       vars changed from USHORT to int, one renamed to reflect       int type.     + Added var names to a few functions declarations.     % Minor optimizations.     * hb_gtScrollEx() modified to not do parameter evaluation,       just pass them to low-level.     ; Przemek, please review me, I've again encountered UCHARs       in hbct / SETCLEARB().
  * source/rtl/idle.c
    * SHORT -> int.
    * Minor formatting.

+++ harbour/contrib/hbct/screen1.c
@@ -312,7 +312,7 @@ static int hb_ctGetClearChar( int iParam )
    else if( HB_ISCHAR( iParam ) )
       iChar = ( UCHAR ) hb_parc( iParam )[0];
    else
-      iChar = hb_gtGetClearChar();
+      iChar = ( int ) hb_gtGetClearChar();
 
    return iChar;
 }
